パッケージ構造の整理: sitestatus -> site/status
