+++ site/status/read_request.go
@@ -12,16 +12,16 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package sitestatus
+package status
 
-import objectstorage "github.com/sacloud/object-storage-api-go"
+import (
+	"github.com/sacloud/packages-go/validate"
+)
 
-// Service provides a high-level API of for Site
-type Service struct {
-	client *objectstorage.Client
+type ReadRequest struct {
+	Id string `service:"-" validate:"required"`
 }
 
-// New returns new service instance of Archive
-func New(client *objectstorage.Client) *Service {
-	return &Service{client: client}
+func (req *ReadRequest) Validate() error {
+	return validate.New().Struct(req)
 }
